+++ packages/@amazeelabs/silverback-gatsby/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.7.15](https://github.com/amazeeio-solutions/silverback-template/compare/@amazeelabs/silverback-gatsby@3.7.14...@amazeelabs/silverback-gatsby@3.7.15) (2026-04-20)
+
+
+### Bug Fixes
+
+* keep language context during path validation in FetchEntity ([b361c8f](https://github.com/amazeeio-solutions/silverback-template/commit/b361c8f3b1baf2aec2e2532180ea3d76a07124fb))
+
+
+
+
+
+## [3.7.14](https://github.com/amazeeio-solutions/silverback-template/compare/@amazeelabs/silverback-gatsby@3.7.13...@amazeelabs/silverback-gatsby@3.7.14) (2026-04-08)
+
+
+### Bug Fixes
+
+* avoid errors in fetch_entity data producer ([ea019e6](https://github.com/amazeeio-solutions/silverback-template/commit/ea019e6a29fc92afa57ed366b87553fe8e456856))
+
+
+
+
+
 ## 3.7.13 (2026-01-22)